added signal emitter
